More careful convenience assigment with Int-ish.

+++ Classes/UInt256+ForwardIndex.swift
@@ -9,9 +9,9 @@
 // Conforming to ForwardIndex results in a duplicate symbol error.
 extension UInt256 { //: ForwardIndex {
     func succ() -> UInt256 {
-        return self + UInt256(1)
+        return self + UInt256(UInt32(1))
     }
     func pred() -> UInt256 {
-        return self + UInt256(1)
+        return self + UInt256(UInt32(1))
     }
 }
